+++ core/eth/reader.go
@@ -710,7 +710,7 @@ func (t *Reader) GetOperatorSocket(ctx context.Context, operatorId core.Operator
 	return socket, nil
 }
 
-func (t *Reader) GetRelayURL(ctx context.Context, key uint16) (string, error) {
+func (t *Reader) GetRelayURL(ctx context.Context, key uint32) (string, error) {
 	if t.bindings.RelayRegistry == nil {
 		return "", errors.New("relay registry not deployed")
 	}
@@ -720,13 +720,13 @@ func (t *Reader) GetRelayURL(ctx context.Context, key uint16) (string, error) {
 	}, uint32(key))
 }
 
-func (t *Reader) GetRelayURLs(ctx context.Context) (map[uint16]string, error) {
+func (t *Reader) GetRelayURLs(ctx context.Context) (map[uint32]string, error) {
 	if t.bindings.RelayRegistry == nil {
 		return nil, errors.New("relay registry not deployed")
 	}
 
-	res := make(map[uint16]string)
-	relayKey := uint16(0)
+	res := make(map[uint32]string)
+	relayKey := uint32(0)
 	for {
 		url, err := t.bindings.RelayRegistry.GetRelayURL(&bind.CallOpts{
 			Context: ctx,
